refactor: extract sync processor parts into smaller files
