+++ lib/features/sync/matrix/sync_event_processor.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: one_member_abstracts, avoid_setters_without_getters
+// ignore_for_file: avoid_setters_without_getters
 
 import 'dart:async';
 import 'dart:collection';
@@ -26,7 +26,8 @@ import 'package:lotti/features/settings/constants/theming_settings_keys.dart';
 import 'package:lotti/features/sync/backfill/backfill_response_handler.dart';
 import 'package:lotti/features/sync/matrix/outbox_bundle_unpacker.dart';
 import 'package:lotti/features/sync/matrix/pipeline/attachment_index.dart';
-import 'package:lotti/features/sync/matrix/utils/atomic_write.dart';
+import 'package:lotti/features/sync/matrix/smart_journal_entity_loader.dart';
+import 'package:lotti/features/sync/matrix/sync_journal_entity_loader.dart';
 import 'package:lotti/features/sync/matrix/utils/attachment_decoding.dart';
 import 'package:lotti/features/sync/model/sync_message.dart';
 import 'package:lotti/features/sync/sequence/sync_sequence_log_service.dart';
@@ -37,496 +38,13 @@ import 'package:lotti/services/db_notification.dart';
 import 'package:lotti/services/domain_logging.dart';
 import 'package:lotti/services/logging_service.dart';
 import 'package:lotti/services/vector_clock_service.dart';
-import 'package:lotti/utils/audio_utils.dart';
 import 'package:lotti/utils/file_utils.dart';
-import 'package:lotti/utils/image_utils.dart';
 import 'package:matrix/matrix.dart';
-import 'package:path/path.dart' as path;
 
-enum VectorClockDecision {
-  accept,
-  retryAfterPurge,
-  staleAfterRefresh,
-  circuitBreaker,
-  missingVectorClock,
-}
-
-class VectorClockValidator {
-  VectorClockValidator({required LoggingService loggingService})
-    : _logging = loggingService;
-
-  static const int maxStaleDescriptorFailures = 5;
-
-  final LoggingService _logging;
-  final Map<String, int> _staleDescriptorFailures = <String, int>{};
-
-  VectorClockDecision evaluate({
-    required String jsonPath,
-    required VectorClock incomingVectorClock,
-    required JournalEntity candidate,
-    required int attempt,
-  }) {
-    final candidateVc = candidate.meta.vectorClock;
-    if (candidateVc == null) {
-      _logging.captureEvent(
-        'smart.fetch.missing_vc path=$jsonPath expected=$incomingVectorClock',
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SmartLoader.fetch',
-      );
-      reset(jsonPath);
-      return VectorClockDecision.missingVectorClock;
-    }
-
-    final status = VectorClock.compare(candidateVc, incomingVectorClock);
-    if (status == VclockStatus.b_gt_a) {
-      final failures = (_staleDescriptorFailures[jsonPath] ?? 0) + 1;
-      _staleDescriptorFailures[jsonPath] = failures;
-      if (failures >= maxStaleDescriptorFailures) {
-        _logging.captureEvent(
-          'smart.fetch.stale_vc.breaker path=$jsonPath retries=$failures limit=$maxStaleDescriptorFailures',
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'SmartLoader.fetch',
-        );
-        return VectorClockDecision.circuitBreaker;
-      }
-      if (attempt == 0) {
-        _logging.captureEvent(
-          'smart.fetch.stale_vc path=$jsonPath expected=$incomingVectorClock got=$candidateVc',
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'SmartLoader.fetch',
-        );
-        return VectorClockDecision.retryAfterPurge;
-      }
-      _logging.captureEvent(
-        'smart.fetch.stale_vc.pending path=$jsonPath expected=$incomingVectorClock got=$candidateVc',
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SmartLoader.fetch',
-      );
-      return VectorClockDecision.staleAfterRefresh;
-    }
-
-    reset(jsonPath);
-    return VectorClockDecision.accept;
-  }
-
-  void reset(String jsonPath) {
-    _staleDescriptorFailures.remove(jsonPath);
-  }
-}
-
-class DescriptorDownloadResult {
-  const DescriptorDownloadResult({
-    required this.json,
-    required this.bytesLength,
-  });
-
-  final String json;
-  final int bytesLength;
-}
-
-class DescriptorDownloader {
-  DescriptorDownloader({
-    required LoggingService loggingService,
-    required VectorClockValidator validator,
-    this.onCachePurge,
-  }) : _logging = loggingService,
-       _validator = validator;
-
-  static const int maxDescriptorDownloadAttempts = 2;
-
-  final LoggingService _logging;
-  final VectorClockValidator _validator;
-  void Function()? onCachePurge;
-
-  Future<DescriptorDownloadResult> download({
-    required Event descriptorEvent,
-    required VectorClock incomingVectorClock,
-    required String jsonPath,
-  }) async {
-    for (var attempt = 0; attempt < maxDescriptorDownloadAttempts; attempt++) {
-      final matrixFile = await downloadAttachmentWithTimeout(
-        descriptorEvent,
-        pathForError: jsonPath,
-      );
-      final downloadedBytes = matrixFile.bytes;
-      if (downloadedBytes.isEmpty) {
-        throw const FileSystemException('empty attachment bytes');
-      }
-      final bytes = await decodeAttachmentBytes(
-        event: descriptorEvent,
-        downloadedBytes: downloadedBytes,
-        relativePath: jsonPath,
-        logging: _logging,
-      );
-      final candidateJson = utf8.decode(bytes);
-      final decoded = json.decode(candidateJson) as Map<String, dynamic>;
-      final candidate = JournalEntity.fromJson(decoded);
-      final decision = _validator.evaluate(
-        jsonPath: jsonPath,
-        incomingVectorClock: incomingVectorClock,
-        candidate: candidate,
-        attempt: attempt,
-      );
-      switch (decision) {
-        case VectorClockDecision.accept:
-          _validator.reset(jsonPath);
-          return DescriptorDownloadResult(
-            json: candidateJson,
-            bytesLength: bytes.length,
-          );
-        case VectorClockDecision.retryAfterPurge:
-          final purged = await _maybePurgeCachedDescriptor(
-            descriptorEvent,
-            jsonPath,
-          );
-          if (purged) {
-            onCachePurge?.call();
-          }
-          _logging.captureEvent(
-            'smart.fetch.stale_vc.refresh path=$jsonPath',
-            domain: 'MATRIX_SERVICE',
-            subDomain: 'SmartLoader.fetch',
-          );
-          continue;
-        case VectorClockDecision.staleAfterRefresh:
-          throw const FileSystemException(
-            'stale attachment json after refresh',
-          );
-        case VectorClockDecision.circuitBreaker:
-          throw const FileSystemException(
-            'stale attachment json (circuit breaker)',
-          );
-        case VectorClockDecision.missingVectorClock:
-          throw const FileSystemException('missing attachment vector clock');
-      }
-    }
-
-    throw const FileSystemException('stale attachment json');
-  }
-
-  Future<bool> _maybePurgeCachedDescriptor(
-    Event event,
-    String jsonPath,
-  ) async {
-    try {
-      final uri = event.attachmentOrThumbnailMxcUrl();
-      if (uri == null) {
-        _logging.captureEvent(
-          'smart.fetch.stale_vc.purge.skipped path=$jsonPath reason=no_mxc',
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'SmartLoader.fetch',
-        );
-        return false;
-      }
-      await event.room.client.database.deleteFile(uri);
-      _logging.captureEvent(
-        'smart.fetch.stale_vc.purge path=$jsonPath mxc=$uri',
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SmartLoader.fetch',
-      );
-      return true;
-    } catch (e, st) {
-      _logging.captureException(
-        e,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SmartLoader.purge',
-        stackTrace: st,
-      );
-      return false;
-    }
-  }
-}
-
-/// Abstraction for loading journal entities and related attachments when
-/// processing sync messages.
-abstract class SyncJournalEntityLoader {
-  Future<JournalEntity> load({
-    required String jsonPath,
-    VectorClock? incomingVectorClock,
-  });
-}
-
-/// Loads journal entities from the documents directory on disk.
-class FileSyncJournalEntityLoader implements SyncJournalEntityLoader {
-  const FileSyncJournalEntityLoader();
-
-  @override
-  Future<JournalEntity> load({
-    required String jsonPath,
-    VectorClock? incomingVectorClock,
-  }) async {
-    final candidateFile = resolveJsonCandidateFile(jsonPath);
-    final docPath = path.normalize(getDocumentsDirectory().path);
-    final jsonRelative = path.relative(candidateFile.path, from: docPath);
-    return readEntityFromJson(jsonRelative);
-  }
-}
-
-/// Smart loader that ensures JSON presence/currency based on an incoming vector
-/// clock. It uses the AttachmentIndex to fetch missing/stale JSON and writes
-/// atomically before parsing.
-class SmartJournalEntityLoader implements SyncJournalEntityLoader {
-  SmartJournalEntityLoader({
-    required AttachmentIndex attachmentIndex,
-    required LoggingService loggingService,
-    void Function()? onCachePurge,
-  }) : _attachmentIndex = attachmentIndex,
-       _logging = loggingService {
-    _vectorClockValidator = VectorClockValidator(
-      loggingService: loggingService,
-    );
-    _descriptorDownloader = DescriptorDownloader(
-      loggingService: loggingService,
-      validator: _vectorClockValidator,
-      onCachePurge: onCachePurge,
-    );
-  }
-
-  final AttachmentIndex _attachmentIndex;
-  final LoggingService _logging;
-  late final VectorClockValidator _vectorClockValidator;
-  late final DescriptorDownloader _descriptorDownloader;
-
-  set onCachePurge(void Function()? listener) {
-    _descriptorDownloader.onCachePurge = listener;
-  }
-
-  void Function(String path)? _onMissingDescriptorPath;
-
-  set onMissingDescriptorPath(void Function(String path)? listener) {
-    _onMissingDescriptorPath = listener;
-  }
-
-  @override
-  Future<JournalEntity> load({
-    required String jsonPath,
-    VectorClock? incomingVectorClock,
-  }) async {
-    final targetFile = resolveJsonCandidateFile(jsonPath);
-    // Build a canonical index key once and reuse it to avoid inconsistencies
-    // across platforms (Windows vs. POSIX). The key always uses forward slashes
-    // and has a single leading '/'. Any leading '/' or '\\' characters are trimmed.
-    final indexKey = _buildIndexKey(jsonPath);
-    // If we have an incoming vector clock, decide whether a fetch is needed.
-    if (incomingVectorClock != null) {
-      try {
-        final local = await const FileSyncJournalEntityLoader().load(
-          jsonPath: jsonPath,
-        );
-        final localVc = local.meta.vectorClock;
-        if (localVc != null) {
-          final status = VectorClock.compare(localVc, incomingVectorClock);
-          if (status == VclockStatus.a_gt_b || status == VclockStatus.equal) {
-            return local; // local is current or newer; no fetch
-          }
-        }
-      } on FileSystemException {
-        // Expected when text arrives before the descriptor: treat as a local miss.
-        // We fetch via AttachmentIndex below; log as an info event rather than exception.
-        _logging.captureEvent(
-          'smart.local.miss path=$jsonPath',
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'SmartLoader.localMiss',
-        );
-      } catch (e, st) {
-        // Unexpected read failure – keep as exception for diagnostics.
-        _logging.captureException(
-          e,
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'SmartLoader.localRead',
-          stackTrace: st,
-        );
-      }
-
-      // Resolve descriptor via AttachmentIndex
-      final eventForPath = _attachmentIndex.find(indexKey);
-      if (eventForPath == null) {
-        // Descriptor not yet available; let caller retry later.
-        _logging.captureEvent(
-          'smart.fetch.miss path=$jsonPath key=$indexKey',
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'SmartLoader.fetch',
-        );
-        throw FileSystemException(
-          'attachment descriptor not yet available',
-          jsonPath,
-        );
-      }
-      try {
-        final descriptor = await _descriptorDownloader.download(
-          descriptorEvent: eventForPath,
-          incomingVectorClock: incomingVectorClock,
-          jsonPath: jsonPath,
-        );
-        await saveJson(targetFile.path, descriptor.json);
-        _logging.captureEvent(
-          'smart.json.written path=$jsonPath bytes=${descriptor.bytesLength}',
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'SmartLoader.fetch',
-        );
-      } catch (e, st) {
-        _logging.captureException(
-          e,
-          domain: 'MATRIX_SERVICE',
-          subDomain: 'SmartLoader.fetchJson',
-          stackTrace: st,
-        );
-        rethrow;
-      }
-    } else {
-      // No incoming vector clock: fetch if file is missing/empty, else read.
-      var needsFetch = false;
-      try {
-        // ignore: avoid_slow_async_io
-        if (!await targetFile.exists()) {
-          needsFetch = true;
-        } else {
-          final len = await targetFile.length();
-          needsFetch = len == 0;
-        }
-      } catch (_) {
-        needsFetch = true;
-      }
-      if (needsFetch) {
-        final eventForPath = _attachmentIndex.find(indexKey);
-        if (eventForPath == null) {
-          _logging.captureEvent(
-            'smart.fetch.miss(noVc) path=$jsonPath key=$indexKey',
-            domain: 'MATRIX_SERVICE',
-            subDomain: 'SmartLoader.fetch',
-          );
-          throw FileSystemException(
-            'attachment descriptor not yet available',
-            jsonPath,
-          );
-        }
-        try {
-          final matrixFile = await downloadAttachmentWithTimeout(
-            eventForPath,
-            pathForError: jsonPath,
-          );
-          final downloadedBytes = matrixFile.bytes;
-          if (downloadedBytes.isEmpty) {
-            throw const FileSystemException('empty attachment bytes');
-          }
-          final bytes = await decodeAttachmentBytes(
-            event: eventForPath,
-            downloadedBytes: downloadedBytes,
-            relativePath: jsonPath,
-            logging: _logging,
-          );
-          final jsonString = utf8.decode(bytes);
-          await saveJson(targetFile.path, jsonString);
-          _logging.captureEvent(
-            'smart.json.written(noVc) path=$jsonPath bytes=${bytes.length}',
-            domain: 'MATRIX_SERVICE',
-            subDomain: 'SmartLoader.fetch',
-          );
-        } catch (e, st) {
-          _logging.captureException(
-            e,
-            domain: 'MATRIX_SERVICE',
-            subDomain: 'SmartLoader.fetchJson.noVc',
-            stackTrace: st,
-          );
-          rethrow;
-        }
-      }
-    }
-
-    // Read and return the entity from disk (either pre-existing or freshly written).
-    final entity = await const FileSyncJournalEntityLoader().load(
-      jsonPath: jsonPath,
-    );
-
-    // Missing media must not block entity application. The JSON payload is the
-    // authoritative state; referenced audio/image files can arrive later via
-    // the attachment pipeline or descriptor catch-up.
-    await _ensureMediaOnMissing(entity);
-    return entity;
-  }
-
-  String _buildIndexKey(String rawPath) => normalizeAttachmentIndexKey(rawPath);
-
-  Future<void> _ensureMediaOnMissing(JournalEntity e) async {
-    switch (e) {
-      case JournalImage():
-        await _ensureMediaFile(getRelativeImagePath(e), mediaType: 'image');
-      case JournalAudio():
-        await _ensureMediaFile(
-          AudioUtils.getRelativeAudioPath(e),
-          mediaType: 'audio',
-        );
-      default:
-        return; // No media to ensure
-    }
-  }
-
-  Future<void> _ensureMediaFile(
-    String relativePath, {
-    String? mediaType,
-  }) async {
-    final docDir = getDocumentsDirectory();
-    final rp = relativePath;
-    // Trim any leading '/' or '\\' to avoid accidental absolute paths on Windows.
-    final rpRel = rp.replaceFirst(RegExp(r'^[\\/]+'), '');
-    final fp = path.normalize(path.join(docDir.path, rpRel));
-    final f = File(fp);
-    try {
-      // ignore: avoid_slow_async_io
-      if (await f.exists()) {
-        final len = await f.length();
-        if (len > 0) return; // present
-      }
-    } catch (e, st) {
-      _logging.captureException(
-        e,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SmartLoader.pathCheck',
-        stackTrace: st,
-      );
-    }
-
-    final descriptorKey = _buildIndexKey(rp);
-    final ev = _attachmentIndex.find(descriptorKey);
-    if (ev == null) {
-      _logging.captureEvent(
-        'smart.media.miss path=$rp key=$descriptorKey',
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SmartLoader.fetchMedia',
-      );
-      _onMissingDescriptorPath?.call(descriptorKey);
-      return;
-    }
-    try {
-      final file = await downloadAttachmentWithTimeout(ev, pathForError: rp);
-      final downloadedBytes = file.bytes;
-      if (downloadedBytes.isEmpty) {
-        throw const FileSystemException('empty attachment bytes');
-      }
-      final bytes = await decodeAttachmentBytes(
-        event: ev,
-        downloadedBytes: downloadedBytes,
-        relativePath: rp,
-        logging: _logging,
-      );
-      await atomicWriteBytes(
-        bytes: bytes,
-        filePath: fp,
-        logging: _logging,
-        subDomain: 'SmartLoader.writeMedia',
-      );
-    } catch (e, st) {
-      _logging.captureException(
-        e,
-        domain: 'MATRIX_SERVICE',
-        subDomain: 'SmartLoader.fetchMedia',
-        stackTrace: st,
-      );
-      _onMissingDescriptorPath?.call(descriptorKey);
-    }
-  }
-}
+export 'package:lotti/features/sync/matrix/descriptor_downloader.dart';
+export 'package:lotti/features/sync/matrix/smart_journal_entity_loader.dart';
+export 'package:lotti/features/sync/matrix/sync_journal_entity_loader.dart';
+export 'package:lotti/features/sync/matrix/vector_clock_validator.dart';
 
 /// Sync message bodies below this base64 length are decoded inline; anything
 /// longer hands off to a worker isolate via `compute`. The break-even point
